requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Bayesian Modeling
pymc3>=3.11.0
arviz>=0.11.0

# Deep Learning
tensorflow>=2.6.0,<2.11.0  # TensorFlow 2.11+ requires Python 3.11
tensorflow-probability>=0.14.0,<0.20.0  # Compatible with TF <2.11

# Causal Inference
dowhy>=0.6.0
econml>=0.12.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Dashboard
fastapi>=0.68.0
uvicorn>=0.15.0
strawberry-graphql>=0.69.0

# Quantum (Optional - Downgraded for Python 3.10)
cirq>=0.13.0,<1.0.0  # cirq 1.x requires Python 3.11+
# tensorflow-quantum removed (no Python 3.10 support on Windows)

# Dev Tools
jupyter>=1.0.0
black>=21.0
flake8>=3.9.0
pytest>=6.2.0